Introduce  GroupAllEventsPanel
