+++ lib/ui/groups/GroupWidgets.dart
@@ -14,12 +14,22 @@
  * limitations under the License.
  */
 
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:illinois/model/Event.dart';
+import 'package:illinois/model/Groups.dart';
 import 'package:illinois/service/Analytics.dart';
+import 'package:illinois/service/AppDateTime.dart';
+import 'package:illinois/service/Groups.dart';
 import 'package:illinois/service/Localization.dart';
 import 'package:illinois/service/Styles.dart';
+import 'package:illinois/ui/events/CreateEventPanel.dart';
+import 'package:illinois/ui/groups/GroupCreatePostPanel.dart';
+import 'package:illinois/ui/groups/GroupsEventDetailPanel.dart';
+import 'package:illinois/ui/widgets/RibbonButton.dart';
 import 'package:illinois/ui/widgets/RoundedButton.dart';
 import 'package:illinois/ui/widgets/ScalableWidgets.dart';
+import 'package:illinois/utils/Utils.dart';
 
 /////////////////////////////////////
 // GroupDropDownButton
@@ -311,5 +321,305 @@ class GroupsConfirmationDialog extends StatelessWidget{
             );
           }),
     );
+  }
+}
+
+//////////////////////////
+//GroupEventCard
+
+class GroupEventCard extends StatefulWidget {
+  final GroupEvent groupEvent;
+  final Group group;
+  final bool isAdmin;
+
+  GroupEventCard({this.groupEvent, this.group, this.isAdmin = false});
+
+  @override
+  createState()=> _GroupEventCardState();
+}
+class _GroupEventCardState extends State<GroupEventCard>{
+  bool _showAllComments = false;
+
+  @override
+  Widget build(BuildContext context) {
+    GroupEvent event = widget.groupEvent;
+    List<Widget> content = [
+      _EventContent(event: event, isAdmin: widget.isAdmin, group: widget.group,),
+    ];
+    List<Widget> content2 = [];
+
+    if(widget.isAdmin){
+      content2.add(
+          Container(
+              padding: EdgeInsets.symmetric(vertical: 16),
+              child:_buildAddPostButton(photoUrl: Groups().getUserMembership(widget.group?.id)?.photoURL,
+                  onTap: (){
+                    Analytics().logPage(name: "Add post");
+                    Navigator.push(context, CupertinoPageRoute(builder: (context) => GroupCreatePostPanel(groupEvent: widget.groupEvent,groupId: widget.group?.id,)));
+                  }))
+      );
+    }
+
+    if (0 < (event?.comments?.length ?? 0)) {
+      content.add(Container(color: Styles().colors.surfaceAccent, height: 1,));
+
+      for (GroupEventComment comment in event.comments) {
+        content2.add(_buildComment(comment));
+        if(!_showAllComments){
+          break;
+        }
+      }
+      if(!_showAllComments && (1 < (event?.comments?.length ?? 0))){
+        content2.add(
+            Container(color: Styles().colors.fillColorSecondary,height: 1,margin: EdgeInsets.only(top:12, bottom: 10),)
+        );
+        content2.add(
+            Semantics(
+              button: true,
+              child: GestureDetector(
+                onTap: (){
+                  setState(() {
+                    _showAllComments = true;
+                  });},
+                child: Center(child: Row(
+                  mainAxisAlignment: MainAxisAlignment.center,
+                  crossAxisAlignment: CrossAxisAlignment.center,
+                  children: <Widget>[
+                    Text(Localization().getStringEx("panel.group_detail.button.previous_post.title", "See previous posts"), style: TextStyle(fontSize: 16,
+                        fontFamily: Styles().fontFamilies.bold,
+                        color: Styles().colors.fillColorPrimary),),
+                    Padding(
+                      padding: EdgeInsets.only(left: 7), child: Image.asset('images/icon-down-orange.png', color:  Styles().colors.fillColorPrimary,),),
+                  ],),
+                ),),
+            )
+        );
+        content2.add(Container(height: 7,));
+      }
+
+      content.add(Padding(padding: EdgeInsets.symmetric(vertical: 8, horizontal: 16), child:
+      Column(crossAxisAlignment: CrossAxisAlignment.start, children:content2))
+      );
+
+    }
+
+    return Padding(
+      padding: const EdgeInsets.symmetric(vertical: 8),
+      child: Container(
+        decoration: BoxDecoration(
+            color: Styles().colors.white,
+            boxShadow: [BoxShadow(color: Styles().colors.blackTransparent018, spreadRadius: 2.0, blurRadius: 6.0, offset: Offset(2, 2))],
+            borderRadius: BorderRadius.all(Radius.circular(8))
+        ),
+        child: Column(crossAxisAlignment: CrossAxisAlignment.start, children: content,),
+      ),
+    );
+  }
+
+  Widget _buildComment(GroupEventComment comment){
+    String memberName = comment.member.name;
+    String postDate = AppDateTime.timeAgoSinceDate(comment.dateCreated);
+    return
+      Semantics(
+          label: "$memberName posted, $postDate: ${comment.text}",
+          excludeSemantics: true,
+          child:Padding(
+              padding: EdgeInsets.symmetric(vertical: 4),
+              child: Container(
+                padding: EdgeInsets.all(15),
+                decoration: BoxDecoration(
+                    color: Styles().colors.white,
+                    boxShadow: [BoxShadow(color: Styles().colors.blackTransparent018, spreadRadius: 1.0, blurRadius: 3.0, offset: Offset(1, 1))],
+                    borderRadius: BorderRadius.all(Radius.circular(4))
+                ),
+                child: Column(crossAxisAlignment: CrossAxisAlignment.start, children: <Widget>[
+                  Row(crossAxisAlignment: CrossAxisAlignment.start, children: <Widget>[
+                    Container(height: 32, width: 32,
+                      decoration: AppString.isStringNotEmpty(comment?.member?.photoURL)
+                          ? BoxDecoration(
+                          shape: BoxShape.circle,
+                          image: DecorationImage(image:NetworkImage(comment.member.photoURL), fit: BoxFit.cover))
+                          : null,
+                    ),
+                    Expanded(
+                      flex: 5,
+                      child: Padding(padding:EdgeInsets.only(left: 8) , child: Column(crossAxisAlignment: CrossAxisAlignment.start, children: <Widget>[
+                        Padding(padding: EdgeInsets.only(bottom: 2), child:
+                        Text(comment.member.name , style: TextStyle(fontFamily: Styles().fontFamilies.bold, fontSize: 14, color: Styles().colors.fillColorPrimary),),
+                        ),
+                        Text(postDate, style: TextStyle(fontFamily: Styles().fontFamilies.regular, fontSize: 12, color: Styles().colors.textBackground),)
+                      ],),),),
+                  ],),
+                  Padding(padding: EdgeInsets.only(top:8), child:
+                  Text(comment.text, style: TextStyle(fontFamily: Styles().fontFamilies.regular, fontSize: 16, color: Styles().colors.textBackground),)
+                  ),
+                ],),
+              )));
+  }
+
+  Widget _buildAddPostButton({String photoUrl,Function onTap}){
+    return
+      InkWell(
+          onTap: onTap,
+          child: Container(
+            child: Row(crossAxisAlignment: CrossAxisAlignment.start, children: <Widget>[
+              photoUrl == null ? Container():
+              Container(height: 32, width: 32,
+                decoration: BoxDecoration(
+                  shape: BoxShape.circle,
+                  image: DecorationImage(image:NetworkImage(photoUrl), fit: BoxFit.cover),
+                ),
+              ),
+              Container(width: 6,),
+              Expanded(child:
+              Container(
+                  height: 45,
+                  alignment: Alignment.centerLeft,
+                  padding:EdgeInsets.symmetric(horizontal: 12) ,
+                  decoration: BoxDecoration(
+                      color: Styles().colors.white,
+                      boxShadow: [BoxShadow(color: Styles().colors.blackTransparent018, spreadRadius: 2.0, blurRadius: 6.0, offset: Offset(2, 2))],
+                      borderRadius: BorderRadius.all(Radius.circular(8))
+                  ),
+                  child:
+                  Text(Localization().getStringEx("panel.group_detail.button.add_post.title", "Add a public post ..."),style: TextStyle(fontSize: 16, color: Styles().colors.textSurface, fontFamily: Styles().fontFamilies.regular),)
+              ))
+            ],),
+          ));
+  }
+}
+
+class _EventContent extends StatelessWidget {
+  final Group group;
+  final Event event;
+  final bool isAdmin;
+
+  _EventContent({this.event, this.isAdmin = false, this.group});
+
+  @override
+  Widget build(BuildContext context) {
+
+    List<Widget> content = [
+      Padding(padding: EdgeInsets.only(bottom: 8, right: 48), child:
+      Text(event?.title ?? '',  style: TextStyle(fontFamily: Styles().fontFamilies.extraBold, fontSize: 20, color: Styles().colors.fillColorPrimary),),
+      ),
+    ];
+    content.add(Padding(padding: EdgeInsets.symmetric(vertical: 4), child: Row(children: <Widget>[
+      Padding(padding: EdgeInsets.only(right: 8), child: Image.asset('images/icon-calendar.png'),),
+      Expanded(child:
+      Text(event.timeDisplayString,  style: TextStyle(fontFamily: Styles().fontFamilies.regular, fontSize: 14, color: Styles().colors.textBackground),)
+      ),
+    ],)),);
+
+    return Stack(children: <Widget>[
+      GestureDetector(onTap: () {
+        Analytics().logPage(name: "Group Settings");
+        Navigator.push(context, CupertinoPageRoute(builder: (context) => GroupEventDetailPanel(event: event,)));
+      },
+          child: Padding(padding: EdgeInsets.all(16), child:
+          Column(crossAxisAlignment: CrossAxisAlignment.start, children: content),
+          )
+      ),
+      Align(alignment: Alignment.topRight,
+          child:
+          Container(
+            padding: EdgeInsets.only(top: 16, right: 16),
+            child: Row(
+              mainAxisAlignment: MainAxisAlignment.end,
+              children: [
+                Semantics(
+                    label: Localization().getStringEx("panel.group_detail.button.favorites.title", "Favorites"),
+                    button: true,
+                    child: GestureDetector(onTap: () { /*TBD switch favorite */ },
+                      child: Container(
+                        child: Image.asset('images/icon-star.png', excludeFromSemantics: true,),
+                      ),
+                    )),
+                !isAdmin? Container() :
+                Container(
+                  padding: EdgeInsets.only(left: 12),
+                  child:
+                  GestureDetector(onTap: () { _onOptionsTap(context);},
+                    child: Container(
+                      child: Image.asset('images/icon-groups-options-orange.png'),
+                    ),
+                  ),
+                )
+              ],),))
+    ],);
+  }
+
+  void _onOptionsTap(BuildContext context){
+    showModalBottomSheet(
+        context: context,
+        backgroundColor: Colors.white,
+        isScrollControlled: true,
+        isDismissible: true,
+        shape: RoundedRectangleBorder(
+          borderRadius: BorderRadius.vertical(top: Radius.circular(24)),
+        ),
+        builder: (context){
+          return Container(
+            padding: EdgeInsets.symmetric(horizontal: 16,vertical: 17),
+            child: Column(
+              mainAxisSize: MainAxisSize.min,
+              children: <Widget>[
+                Container(height: 48,),
+                RibbonButton(
+                  height: null,
+                  leftIcon: "images/icon-leave-group.png",
+                  label:Localization().getStringEx("panel.group_detail.button.remove_event.title", "Remove Event"),
+                  onTap: (){
+                    showDialog(context: context, builder: (context)=>_buildRemoveEventDialog(context)).then((value) => Navigator.pop(context));
+                  },
+                ),
+                RibbonButton(
+                  height: null,
+                  leftIcon: "images/icon-leave-group.png",
+                  label:Localization().getStringEx("panel.group_detail.button.delete_event.title", "Delete Event"),
+                  onTap: (){
+                    showDialog(context: context, builder: (context)=>_buildDeleteEventDialog(context)).then((value) => Navigator.pop(context));
+                  },
+                ),
+                RibbonButton(
+                  height: null,
+                  leftIcon: "images/icon-edit.png",
+                  label:Localization().getStringEx("panel.group_detail.button.edit_event.title", "Edit Event"),
+                  onTap: (){
+                    _onEditEventTap(context);
+                  },
+                ),
+              ],
+            ),
+          );
+        }
+    );
+  }
+
+  Widget _buildRemoveEventDialog(BuildContext context){
+    return GroupsConfirmationDialog(
+        message: Localization().getStringEx("panel.group_detail.button.remove_event.title",  "Remove this event from your group page?"),
+        buttonTitle:Localization().getStringEx("panel.group_detail.button.remove.title", "Remove"),
+        onConfirmTap:_onRemoveEvent);
+  }
+
+  Widget _buildDeleteEventDialog(BuildContext context){
+    return GroupsConfirmationDialog(
+        message: Localization().getStringEx("panel.group_detail.button.delete_event.title", "Delete this event from your groups page?"),
+        buttonTitle:  Localization().getStringEx("panel.group_detail.button.delete.title","Delete"),
+        onConfirmTap:_onDeleteEvent);
+  }
+
+  void _onRemoveEvent(){
+    //TBD
+  }
+
+  void _onDeleteEvent(){
+    //TBD
+  }
+
+  void _onEditEventTap(BuildContext context){
+    Analytics().logPage(name: "Create Event");
+    Navigator.push(context, MaterialPageRoute(builder: (context) => CreateEventPanel(group: group,)));
   }
 }
